+++ projects/swimlane/ngx-charts/src/lib/gauge/percent-gauge/percent-gauge.component.ts
@@ -1,5 +1,4 @@
-import { Component, Input, ViewEncapsulation, ChangeDetectionStrategy, ElementRef, ViewChild } from '@angular/core';
-import { GradientPath } from 'gradient-path';
+import { Component, Input, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/core';
 
 import { BaseChartComponent } from '../../common/base-chart.component';
 import { calculateViewDimensions } from '../../common/view-dimensions.helper';
@@ -26,11 +25,10 @@ import { ScaleType } from '../../common/types/scale-type.enum';
               [style.stroke-dashoffset]="circumference * (1 - percent / 100)"
             />
           </mask>
-          <text x="0" y="0" fill="white" stroke="none" class="total" [style.font-size]="valueFontSize">
+          <text x="0" y="0" stroke="none" class="total" [style.font-size]="valueFontSize">
             {{ displayValue }}
           </text>
           <circle
-            #circleEl
             class="dashes-back"
             [style.stroke-width]="radius / 5"
             [attr.r]="radius"
@@ -56,6 +54,14 @@ import { ScaleType } from '../../common/types/scale-type.enum';
 
           <svg:g [attr.transform]="targetTransform">
             <circle
+              class="target-circle-bg"
+              [attr.r]="targetRadius"
+              [style.stroke-width]="targetRadius / 10"
+              [attr.cx]="-targetRadius / 2"
+              [attr.cy]="-targetRadius / 2"
+            />
+            <circle
+              *ngIf="percent >= target"
               class="target-circle"
               [attr.r]="targetRadius"
               [style.stroke-width]="targetRadius / 10"
@@ -105,8 +111,6 @@ export class PercentGaugeComponent extends BaseChartComponent {
   @Input() label: string;
   @Input() valueFormatting: any;
   @Input() showLabel = true;
-
-  @ViewChild('circleEl') circleElement: ElementRef;
 
   defaultMargin: number[] = [20, 40, 20, 40];
   margin: number[] = [20, 40, 20, 40];
@@ -182,46 +186,34 @@ export class PercentGaugeComponent extends BaseChartComponent {
     this.cd.markForCheck();
   }
 
+  generateCirclePoints(radius: number, numPoints: number): { x: number; y: number }[] {
+    const points = [];
+    for (let i = 0; i < numPoints; i++) {
+      const angle = (i / numPoints) * 2 * Math.PI;
+      const x = radius * Math.cos(angle);
+      const y = radius * Math.sin(angle);
+      points.push({ x, y });
+    }
+    return points;
+  }
+
   generateticks() {
-    if (this.circleElement?.nativeElement) {
-      const clonedCircle = this.circleElement.nativeElement.cloneNode(true);
-      clonedCircle.setAttribute('stroke-width', this.radius / 5);
-      clonedCircle.setAttribute('r', this.radius);
-
-      this.circleElement.nativeElement.parentElement.appendChild(clonedCircle);
-
-      try {
-        const gp = new GradientPath({
-          path: clonedCircle,
-          segments: 60,
-          samples: 2,
-          precision: 2
-        });
-
-        clonedCircle.remove();
-
-        const data = gp.data.flatMap(({ samples }) => samples);
-        this.ticks = [];
-        this.circleTransform = `rotate(-90,0,0)`;
-        for (let j = 0; j < data.length; j++) {
-          const { x, y } = data[j];
-          let progress = data[j].progress;
-          if (progress === 1) {
-            progress = 0;
-          }
-
-          this.ticks.push({
-            height: this.ticHeight,
-            width: this.radius / 60,
-            fill: this.colors.getColor(progress * this.max),
-            transform: `translate(${x}, ${y}), rotate(${360 * progress - 90})`
-          });
-        }
-      } catch (e) {
-        console.log('Unable to generate ticks for percent gauge chart', e);
-      } finally {
-        clonedCircle.remove();
+    const numPoints = 60;
+    const points = this.generateCirclePoints(this.radius, numPoints);
+    this.ticks = [];
+    this.circleTransform = `rotate(-90,0,0)`;
+    for (let j = 0; j < points.length; j++) {
+      const { x, y } = points[j];
+      let progress = j / numPoints;
+      if (progress === 1) {
+        progress = 0;
       }
+      this.ticks.push({
+        height: this.ticHeight,
+        width: this.radius / 60,
+        fill: this.colors.getColor(progress * this.max),
+        transform: `translate(${x}, ${y}), rotate(${360 * progress - 90})`
+      });
     }
   }